Squashed commit of the following:
commit 1e2acc54f4d38e4152ea452b155f249a5ec2a443
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 08:38:24 2018 -0400

    Fix error

commit 8d474585ea790b379bbf5cd56be5c3ed69f77e22
Merge: dde6388d 1c482a62
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 08:36:30 2018 -0400

    Merge branch 'develop' into feature/GlidenArm

    # Conflicts:
    #	Provenance/Settings/PVSettingsViewController.swift
    #	ProvenanceTV/PVTVSettingsViewController.swift

commit dde6388dea2c5670970cc785c5864abaedd1704a
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 08:32:28 2018 -0400

    Fix gliden ini paths

commit f206653ca9311bbc3e8358ab0f582e5129e23b99
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 08:32:16 2018 -0400

    Hard reset of submodule

commit 5d0e4623af3cacb15d52362093cbf2326a2b62e3
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 07:57:57 2018 -0400

    Move gliden to Plugins folder

    I can’t merge without renaming since the folder of the submodule is the same as the old directory

commit 4e18e15911106839adf8ae4312328b272f3b34e6
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 07:40:10 2018 -0400

    Add new logging function, update submodule

commit 79185b9adaf542436a83870f5bc24ace25377713
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 07:01:56 2018 -0400

    Fix tvOS build of gliden

commit 85759b8b8ce0cb424c4e3749fd9aa6069df12a83
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 06:47:46 2018 -0400

    Fix header search and iPhone simulator build

commit f5164bfb8b01ec2ecaf0bf48bee77a75aab8a962
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 05:56:57 2018 -0400

    Update our submodule of gliden

commit 8f81b15ea94214d11e356647ecf5458ad68142aa
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 05:45:13 2018 -0400

    Add our fork submodule for gliden

commit f153543bef1a745bd114af50cfa38799c6710396
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 05:44:22 2018 -0400

    Remove our local copy of gliden to get ready for submodule

commit 662016a71be2a6454e51a15c1c23edcf0274cec1
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 05:36:30 2018 -0400

    Update some gliden64 compile settings

commit 27b2903faa26626daa0591bb8aaf8170a268db35
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 05:23:00 2018 -0400

    iOS Log to the XCode terminal

commit ae669565e0d113c18d435306b421d0f1c33d34c7
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 05:22:40 2018 -0400

    Get rid of NO_FREETYPE by using stub class

commit 5f196524a82b26af0700f641b3af60c872761cd3
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 05:04:09 2018 -0400

    Add back @braindx ios framebuffer binding

commit c1fdaaf9b784417196b92efa5b2b39ffdcdcda47
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 02:22:39 2018 -0400

    Don’t precompile brdige header for Version.h support

commit f5d14a41d03d4dd499a363b87fbb12699b8aa94d
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Fri Mar 30 02:22:08 2018 -0400

    Add back an ios ifdef

commit 3623a9efdc6aa60587917efc123490b5e52cb355
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Thu Mar 29 01:19:26 2018 -0400

    Undo a line that was commented out, not sure why it was

commit 50f57fa9c1ef1f43c5cfc24800fa21649a1dd53e
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Thu Mar 29 00:59:35 2018 -0400

    Update more gliden files

commit 97994fde5d4faafa203da53f281452e70763f237
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Thu Mar 29 00:49:36 2018 -0400

    add texture CRC arm

commit fa6cb5a2a0e18901bb16947c4955d2a7c6db9612
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Thu Mar 29 00:31:37 2018 -0400

    Some reason cxd4 file is bothering gliden

commit 00fca0cb0a536c08cfd78a49bca6e89983f34ae0
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Thu Mar 29 00:29:22 2018 -0400

    Show GIT branch with revision if not built from master branch

    to help see which branch we’re building off of, but should be hidden for releases since they will be on master

commit 75809cb808bd951d46873103dd5e49b1cdad3576
Author: Joe Mattiello <mail@joemattiello.com>
Date:   Thu Mar 29 00:09:26 2018 -0400

    Update Gliden code for new arm64 instrinics

+++ PVMupen64Plus/mupen64plus-rsp-cxd4/my_types.h
@@ -241,8 +241,8 @@ typedef signed __int64          s64;
 typedef unsigned __int64        u64;
 
 #elif defined(__LP64__) && (0x00000000FFFFFFFFUL < ~0UL)
-typedef signed long             s64;
-typedef unsigned long           u64;
+typedef long long             s64;
+typedef unsigned long long           u64;
 #elif (LONG_MIN < -9223372036854775807L && LONG_MAX >= +9223372036854775807L)
 typedef signed long             s64;
 typedef unsigned long           u64;
